emacs: patch cd to stay in remote path
When the path starts with a single '/', this patch keeps the path restricted
to the root of the remote tramp path. To go to actual root, use '//'

This is an experimental feature; I may replace it with a different mechanism.
I am not very happy about other commands not benefiting from the patch.