emacs: add tags to time log entries
I want to check the category of items I spend time on; so I added this
feature that picks up the log tag from subtree or parent, and add that as
a hashtag to timelog. I don't want to overyly complicate this, so I added
support for a single tag.