emacs: add ssl-xoauth2 support for SMTP
My pia email moved to office-365, which promptly broke my mail system.
This change brings it back into the fold. There are some things to be done
by the site administrator; but after that I also had to deal with short
lived access tokens. Which eventually led to this commit.